+++ extaz/availablity_zone_discovery.go
@@ -6,24 +6,49 @@ package extaz
 import (
 	"context"
 	"errors"
+	"fmt"
 	"github.com/aws/aws-sdk-go-v2/aws"
 	awshttp "github.com/aws/aws-sdk-go-v2/aws/transport/http"
 	"github.com/aws/aws-sdk-go-v2/service/ec2"
 	types2 "github.com/aws/aws-sdk-go-v2/service/ec2/types"
 	"github.com/rs/zerolog/log"
 	"github.com/steadybit/discovery-kit/go/discovery_kit_api"
+	"github.com/steadybit/extension-aws/config"
 	"github.com/steadybit/extension-aws/utils"
 	extension_kit "github.com/steadybit/extension-kit"
 	"github.com/steadybit/extension-kit/extbuild"
 	"github.com/steadybit/extension-kit/exthttp"
 	"github.com/steadybit/extension-kit/extutil"
 	"net/http"
+	"time"
+)
+
+var (
+	targets        []discovery_kit_api.Target
+	discoveryError *extension_kit.ExtensionError
 )
 
 func RegisterDiscoveryHandlers() {
 	exthttp.RegisterHttpHandler("/az/discovery", exthttp.GetterAsHandler(getAZDiscoveryDescription))
 	exthttp.RegisterHttpHandler("/az/discovery/target-description", exthttp.GetterAsHandler(getAZTargetDescription))
 	exthttp.RegisterHttpHandler("/az/discovery/discovered-targets", getAZDiscoveryResults)
+	targets = []discovery_kit_api.Target{}
+	go func() {
+		for {
+			start := time.Now()
+			updatedTargets, err := utils.ForEveryAccount(utils.Accounts, getTargetsForAccount, context.Background(), "availability zones")
+			if err != nil {
+				discoveryError = extutil.Ptr(extension_kit.ToError("Failed to collect availability zones", err))
+				targets = []discovery_kit_api.Target{}
+			} else {
+				discoveryError = nil
+				targets = *updatedTargets
+			}
+			elapsed := time.Since(start)
+			log.Debug().Msgf("Updated %d availability zone targets in %s", len(targets), elapsed)
+			time.Sleep(time.Duration(config.Config.DiscoveryIntervalZone) * time.Second)
+		}
+	}()
 }
 
 func getAZDiscoveryDescription() discovery_kit_api.DiscoveryDescription {
@@ -33,7 +58,7 @@ func getAZDiscoveryDescription() discovery_kit_api.DiscoveryDescription {
 		Discover: discovery_kit_api.DescribingEndpointReferenceWithCallInterval{
 			Method:       "GET",
 			Path:         "/az/discovery/discovered-targets",
-			CallInterval: extutil.Ptr("300s"),
+			CallInterval: extutil.Ptr(fmt.Sprintf("%ds", config.Config.DiscoveryIntervalZone)),
 		},
 	}
 }
@@ -62,9 +87,8 @@ func getAZTargetDescription() discovery_kit_api.TargetDescription {
 }
 
 func getAZDiscoveryResults(w http.ResponseWriter, r *http.Request, _ []byte) {
-	targets, err := utils.ForEveryAccount(utils.Accounts, getTargetsForAccount, mergeTargets, make([]discovery_kit_api.Target, 0, 100), r.Context())
-	if err != nil {
-		exthttp.WriteError(w, extension_kit.ToError("Failed to collect availability zones information", err))
+	if discoveryError != nil {
+		exthttp.WriteError(w, *discoveryError)
 	} else {
 		exthttp.WriteBody(w, discovery_kit_api.DiscoveryData{Targets: &targets})
 	}
@@ -77,10 +101,6 @@ func getTargetsForAccount(account *utils.AwsAccount, ctx context.Context) (*[]di
 		return nil, err
 	}
 	return &targets, nil
-}
-
-func mergeTargets(merged []discovery_kit_api.Target, eachResult []discovery_kit_api.Target) ([]discovery_kit_api.Target, error) {
-	return append(merged, eachResult...), nil
 }
 
 type AZDescribeAvailabilityZonesApi interface {